+++ index.html
@@ -28,20 +28,19 @@
                     </div>
                 </div>
             </header>
-            <!-- <img
-                class="mobile-menu-button | img-width-40 hidden"
-                src="/menu_icon.svg"
-                alt="Mobline Menu Icon" /> -->
             <svg
                 class="mobile-menu-button | hidden"
-                xmlns="http://www.w3.org/2000/svg"
-                width="40"
-                height="24.286"
-                viewBox="0 0 40 24.286"
-                fill="">
-                <path
-                    id="menuIcon"
-                    d="M2.143,24.286a2.143,2.143,0,0,1,0-4.286H37.857a2.143,2.143,0,0,1,0,4.286Zm0-10a2.143,2.143,0,0,1,0-4.286H37.857a2.143,2.143,0,0,1,0,4.286Zm0-10A2.143,2.143,0,0,1,2.143,0H37.857a2.143,2.143,0,0,1,0,4.286Z" />
+                data-menu-state="closed"
+                width="42"
+                height="24"
+                stroke=""
+                stroke-width="4"
+                stroke-linecap="round">
+                <path class="line-top hamburger" d="M2 2H40" />
+                <path class="line-mid hamburger" d="M2 12H40" />
+                <path class="x x-1" d="M2 12H40" />
+                <path class="x x-2" d="M2 12H40" />
+                <path class="line-bot hamburger" d="M2 22H40" />
             </svg>
             <nav class="flex-grow size-font-xl" data-menu-state="closed">
                 <ul role="list" class="flex-column gap">
@@ -62,73 +61,5 @@
                 </p>
             </footer>
         </div>
-        <!-- <nav class="navigation-bar">
-            <div class="logo-and-titles-container">
-                <img
-                    type=""
-                    class="site-logo"
-                    src="images/ryskDigitalLogoDark.png"
-                    alt=""
-                />
-                <div>
-                    <h1 class="artist-name">Ryszard Klimczewski</h1>
-                    <h2 class="artist-title">Technical 3D Artist</h2>
-                </div>
-            </div>
-            <div id="navigation-menu" class="nav-menu-links-container">
-                <div
-                    class="close-navigation-button"
-                    onclick="hideNavigationMenu()"
-                >
-                    <p>X</p>
-                </div>
-                <ul>
-                    <li>
-                        <a
-                            class="navigation-link current-page"
-                            href="index.html"
-                            >Home</a
-                        >
-                    </li>
-                    <li>
-                        <a
-                            class="navigation-link"
-                            href="/portfolio-page/portfolio.html"
-                        >
-                            Portfolio
-                        </a>
-                    </li>
-                    <li>
-                        <a class="navigation-link" href="Resume.html">Resume</a>
-                    </li>
-                </ul>
-            </div>
-            <div class="mobile-menu-button">
-                <div class="svg-container" onclick="displayNavigationMenu()">
-                    <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="40"
-                        height="24.286"
-                        viewBox="0 0 40 24.286"
-                    >
-                        <path
-                            id="menuIcon"
-                            d="M2.143,24.286a2.143,2.143,0,0,1,0-4.286H37.857a2.143,2.143,0,0,1,0,4.286Zm0-10a2.143,2.143,0,0,1,0-4.286H37.857a2.143,2.143,0,0,1,0,4.286Zm0-10A2.143,2.143,0,0,1,2.143,0H37.857a2.143,2.143,0,0,1,0,4.286Z"
-                            fill="#979797"
-                        />
-                    </svg>
-                </div>
-            </div>
-            <div class="navigation-menu-footer">
-                <p>&copy; Ryszard Klimczewski, 2021</p>
-            </div>
-        </nav> -->
-        <!-- 
-        <script>
-            setInterval(changeBackgroundImage, 5000);
-        </script>
-        <script>
-            styleCurrentNavLink();
-        </script> -->
     </body>
 </html>
