+++ scss/blocks/_background-container.scss
@@ -8,3 +8,9 @@
     background-position: center center;
     transition: background-image 1s linear;
 }
+
+@media (max-width: 1024px) {
+    .background-container {
+        background-position: 60% center;
+    }
+}
