+++ scss/main.scss
@@ -24,6 +24,10 @@ a {
     color: var(--color-font-base);
 }
 
+a:hover {
+    color: var(--color-font-accent-base);
+}
+
 ul,
 ol {
     margin: 0;
